Added xcodeproj writer support for local library references

+++ spec/integration/xcodeproj_writer_spec.rb
@@ -3,29 +3,6 @@ require 'fastlane'
 
 RSpec.describe StructCore::XcodeprojWriter do
 	describe '#write' do
-		it 'can write a working project with pod references' do
-			next if should_stub_tests_on_incompatible_os
-
-			destination = Dir.mktmpdir
-			copy_support_files File.join(File.dirname(__FILE__), 'support_files', 'xcodeproj_writer_test_pod_references'), destination
-
-			Dir.chdir(destination) do
-				`pod install`
-			end
-
-			spec = StructCore::SpecBuilder.build File.join(destination, 'Specfile')
-
-			expect { StructCore::XcodeprojWriter.write spec, destination }.to_not raise_error
-
-			project_file = File.join destination, 'project.xcodeproj'
-			Fastlane.load_actions
-			Dir.chdir(File.join(File.dirname(__FILE__), 'support_files')) {
-				expect { Fastlane::LaneManager.cruise_lane(nil, 'build', {:project => project_file}, {}) }.to_not raise_error
-			}
-
-			FileUtils.rm_rf destination
-		end
-
 		it 'can write a working project with xcconfig configurations' do
 			next if should_stub_tests_on_incompatible_os
 
@@ -112,6 +89,46 @@ RSpec.describe StructCore::XcodeprojWriter do
 
 			expect { StructCore::XcodeprojWriter.write proj, destination }.to_not raise_error
 			expect(File.exist?(File.join(destination, 'My_variant_project_file.xcodeproj'))).to be_truthy
+
+			FileUtils.rm_rf destination
+		end
+
+		it 'can write a working project that links to local libraries' do
+			next if should_stub_tests_on_incompatible_os
+
+			destination = Dir.mktmpdir
+			copy_support_files File.join(File.dirname(__FILE__), 'support_files', 'xcodeproj_writer_test_local_libraries'), destination
+
+			spec = StructCore::Specparser.new.parse File.join(destination, 'project.yml')
+
+			expect { StructCore::XcodeprojWriter.write spec, destination }.to_not raise_error
+
+			project_file = File.join destination, 'project.xcodeproj'
+			Fastlane.load_actions
+			Dir.chdir(File.join(File.dirname(__FILE__), 'support_files')) {
+				expect { Fastlane::LaneManager.cruise_lane(nil, 'build', {:project => project_file}, {}) }.to_not raise_error
+			}
+		end
+
+		it 'can write a working project with pod references' do
+			next if should_stub_tests_on_incompatible_os
+
+			destination = Dir.mktmpdir
+			copy_support_files File.join(File.dirname(__FILE__), 'support_files', 'xcodeproj_writer_test_pod_references'), destination
+
+			Dir.chdir(destination) do
+				`pod install`
+			end
+
+			spec = StructCore::SpecBuilder.build File.join(destination, 'Specfile')
+
+			expect { StructCore::XcodeprojWriter.write spec, destination }.to_not raise_error
+
+			project_file = File.join destination, 'project.xcodeproj'
+			Fastlane.load_actions
+			Dir.chdir(File.join(File.dirname(__FILE__), 'support_files')) {
+				expect { Fastlane::LaneManager.cruise_lane(nil, 'build', {:project => project_file}, {}) }.to_not raise_error
+			}
 
 			FileUtils.rm_rf destination
 		end
